Soham_Mukherjee/google-agentic-suite/pyproject.toml
[project]
name = "google-agentic-suite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.181.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
]
